feat: implement suggestions from @tusharmath
