+++ crates/forge_domain/src/lib.rs
@@ -5,6 +5,7 @@ mod chat_request;
 mod chat_response;
 mod compaction_result;
 mod conversation_html;
+mod update;
 
 mod context;
 mod conversation;
@@ -74,4 +75,5 @@ pub use tool_definition::*;
 pub use tool_name::*;
 pub use tool_result::*;
 pub use tool_usage::*;
+pub use update::*;
 pub use workflow::*;
