+++ crates/forge_main/src/lib.rs
@@ -11,7 +11,7 @@ mod state;
 mod tools_display;
 mod ui;
 
-pub use auto_update::{check_for_update, UpdateConfiguration, UpdateFrequency};
+pub use auto_update::check_for_update;
 pub use cli::Cli;
 use lazy_static::lazy_static;
 pub use ui::UI;
